test: reorganize disputeValidation tests
