+++ package.json
@@ -84,21 +84,21 @@
     "test:debug": "yarn hardhat test --no-compile",
     "test:all": "yarn hardhat test",
     "test:compile": "yarn compile && yarn test",
-    "test:e2e": "LOG_LEVEL=warn yarn hardhat test test/e2e/*",
+    "test:e2e": "LOG_LEVEL=warn bash -O globstar -c 'yarn hardhat test test/e2e/**/*.test.ts \"$@\"' --",
     "test:e2e:log-file": "rm -rf ./logs/wtf.ansi && yarn test:e2e:verbose >> ./logs/wtf.ansi",
     "test:e2e:skip-log": "LOG_SKIP_WRITING=true yarn test:e2e:verbose",
-    "test:e2e:info": "LOG_LEVEL=info yarn hardhat test test/e2e/*",
-    "test:e2e:debug": "LOG_LEVEL=debug yarn hardhat test test/e2e/*",
-    "test:e2e:verbose": "LOG_LEVEL=verbose yarn hardhat test test/e2e/*",
+    "test:e2e:info": "LOG_LEVEL=info bash -O globstar -c 'yarn hardhat test test/e2e/**/*.test.ts \"$@\"' --",
+    "test:e2e:debug": "LOG_LEVEL=debug bash -O globstar -c 'yarn hardhat test test/e2e/**/*.test.ts \"$@\"' --",
+    "test:e2e:verbose": "LOG_LEVEL=verbose bash -O globstar -c 'yarn hardhat test test/e2e/**/*.test.ts \"$@\"' --",
     "test:e2e:parallel": "node scripts/test-e2e-parallel.js",
     "test:e2e:parallel:debug": "LOG_LEVEL=debug node scripts/test-e2e-parallel.js",
     "test:e2e:parallel:verbose": "LOG_LEVEL=verbose node scripts/test-e2e-parallel.js",
     "test:coverage": "nyc --reporter=html yarn test",
-    "test:profile:e2e": "LOG_LEVEL=verbose node --enable-source-maps --cpu-prof --cpu-prof-dir=./profiles --cpu-prof-name=e2e-verbose.cpuprofile ./node_modules/hardhat/internal/cli/cli.js test test/e2e/*",
-    "test:profile:e2e:inspector": "LOG_LEVEL=verbose node --inspect-brk --enable-source-maps ./node_modules/hardhat/internal/cli/cli.js test test/e2e/*",
-    "test:profile:e2e:clinic:flame": "LOG_LEVEL=verbose npx clinic flame --dest ./profiles/clinic -- node ./node_modules/hardhat/internal/cli/cli.js test test/e2e/*",
-    "test:profile:e2e:clinic:doctor": "LOG_LEVEL=verbose npx clinic doctor --dest ./profiles/clinic -- node ./node_modules/hardhat/internal/cli/cli.js test test/e2e/*",
-    "test:profile:e2e:v8": "LOG_LEVEL=verbose node --prof --enable-source-maps ./node_modules/hardhat/internal/cli/cli.js test test/e2e/*",
+    "test:profile:e2e": "LOG_LEVEL=verbose bash -O globstar -c 'node --enable-source-maps --cpu-prof --cpu-prof-dir=./profiles --cpu-prof-name=e2e-verbose.cpuprofile ./node_modules/hardhat/internal/cli/cli.js test test/e2e/**/*.test.ts \"$@\"' --",
+    "test:profile:e2e:inspector": "LOG_LEVEL=verbose bash -O globstar -c 'node --inspect-brk --enable-source-maps ./node_modules/hardhat/internal/cli/cli.js test test/e2e/**/*.test.ts \"$@\"' --",
+    "test:profile:e2e:clinic:flame": "LOG_LEVEL=verbose bash -O globstar -c 'npx clinic flame --dest ./profiles/clinic -- node ./node_modules/hardhat/internal/cli/cli.js test test/e2e/**/*.test.ts \"$@\"' --",
+    "test:profile:e2e:clinic:doctor": "LOG_LEVEL=verbose bash -O globstar -c 'npx clinic doctor --dest ./profiles/clinic -- node ./node_modules/hardhat/internal/cli/cli.js test test/e2e/**/*.test.ts \"$@\"' --",
+    "test:profile:e2e:v8": "LOG_LEVEL=verbose bash -O globstar -c 'node --prof --enable-source-maps ./node_modules/hardhat/internal/cli/cli.js test test/e2e/**/*.test.ts \"$@\"' --",
     "test:profile:e2e:v8:process": "node --prof-process isolate-*.log > ./profiles/v8-prof.txt",
     "build": "yarn compile && rimraf dist && mkdir -p dist && tsc && tsc-alias -p tsconfig.json && npm pack && node rename-pack.js",
     "format": "prettier --write .",
